Task Runner: Let tasks be "disabled"; choose appropriate default
Summary:
This change lets task runners specify whether each of there tasks is currently enabled or not. The idea is that, for example, if your active project isn't a Buck project, Buck tasks shouldn't be selectable.

In the future, the exact UI for this may change. For example, we may group all disabled tasks at the bottom of the dropdown. I also want to add a `disabledReason` for communicating why a task isn't enabled (T11659298). The important thing is that the Task Runner package should decide how to display these (or not to). So task runners should always report all of their available tasks—even if some of them are disabled.

Reviewed By: hansonw

Differential Revision: D3867689

fbshipit-source-id: 42b079a0289ed008cae993155b96b63eb6106c1a