Use nullable environment rather than isAvailable
Summary:
This next stack cleans up the HackLanguage creation.
Changes it to use the ConnectionCache, rather than caching per hack-root.
Then switches the service to be a class rather than individual functions.

This diff:

Currently getHackEnvironment() returns an object with an isAvailable member,
when isAvailable is false the other members are null. This first simplification
just removes the isAvailable member and returns null if hack is not present.

Reviewed By: mostafaeweda

Differential Revision: D3872706

fbshipit-source-id: 83bc36f2e93791b624c909561c2fb688b9d50e72