Increase quick-open modal margin to accommodate custom title bar traffic light buttons
Summary:
With Atom 1.11, there's now an option to use a custom title bar. Since this will render in the document, the modal will cover it. That's okay, but the traffic lights cover the modal.

As a stop-gap, this decreases the width of the quick open modal so that it won't overlap the traffic lights. Note that normal modals (which we don't change the width of) still have this problem with small enough windows. This should be fixed on the theme side though.

Reviewed By: zertosh

Differential Revision: D3877335

fbshipit-source-id: f868601fb3f6a9447cfe2485517b0000bc684a31