Reorganize Atom opener scripts to make fast editing of a file in Atom easier to set up.
Summary:
The existing `pkg/nuclide-remote-atom-rpc/bin/atom` script is very
useful on a remote machine, but it does not fail gracefully. This renames
it to `atom-rpc` and introduces `pkg/nuclide-remote-atom-rpc/bin/atom`
to provide the option to fall back on a local editor. This makes
`pkg/nuclide-remote-atom-rpc/bin/atom --wait` a particularly good choice
for things like `git config --global core.editor`.

This includes substantial documentation updates in an attempt to make
it as easy as possible for end users to set this up for themselves.

Reviewed By: peterhal

Differential Revision: D3885235

fbshipit-source-id: c342941af054c4677f67a6583281edc2ef3b19ab