Move Hack formatSource to use FileVersion
Summary:
This is the first diff that cleans up an existing Hack language service API.

It does several refactorings which will be repeated for each API:
It uses FileVersion instead of filePath & contents.
It uses atom$Range(or atom$Point) instead of explicit line/col or custom location types.
The resulting API has no hack specific data structures/types in it.
It centralizes all the logic in the HackService where possible.
The HackLanguage is just a simple pass through to the HackService.

Reviewed By: mostafaeweda

Differential Revision: D3873184

fbshipit-source-id: 3d6ea25f27b0c6e3b7081987bc10cf72abd8bf0c