Save form in Remote Connection Profile Builder when enter key is pressed
Summary:
When the Remote Connection profile building dialog, hitting 'enter' with the password input box focussed will not perform the "Save" action -- however, doing so in any other input box inside this form will.

This is because all the other input boxes are actually `AtomInput` React components. It is assumed that since the `componentDidMount` sets up an atom command to handle saving when the enter key is pressed with the dialog focused (https://github.com/facebook/nuclide/blob/master/pkg/nuclide-remote-projects/lib/CreateConnectionProfileForm.js#L71) that this case is handled. However, the input box for the password field is not `AtomInput` but a regular `input` box.

Fix:
Instead of passing in an empty function, pass in the `_clickSave` handler. This function is then called in `ConnectionDetailsForm` at https://github.com/facebook/nuclide/blob/master/pkg/nuclide-remote-projects/lib/ConnectionDetailsForm.js#L85 when the Enter key is pressed.

Before:
https://i.imgur.com/sxi8G7N.gifv

After:
https://i.
Closes https://github.com/facebook/nuclide/pull/714

Differential Revision: D3891822

Pulled By: zertosh

fbshipit-source-id: f32fc720d769a88fe102367c44987a85c31349f7