Hide Find and Replace when switching to distraction-free mode in Nuclide
Summary:
Similarly to task 3145697, the intent of this change is to hide the Find&Replace tool when entering distraction-free mode on Nuclide.
Trying to use a similar approach to 3145697 (use DOM API to detect visibility) and by controlling the tool via commands dispatch I could obtain the desired behavior (I used the file tree pane as reference).

The find&replace pane is actually composed of two features, each with their own set of commands and panes, called project-find and find-and-replace. I used the DOM API to determine which feature is visible when entering DF mode based on a display attribute. There is no :hide command, therefore both features are only controlled by a :toggle and a :show command (in regards to their visibility).

Reviewed By: nmote

Differential Revision: D3887100

fbshipit-source-id: ddd56a91d7b42e8fabae060e1eddd8e63e1d62a9