Fix jumpy header in diagnostics table
Summary: When a table header was clicked, the header title was jumping 1px. This was because the triangle icons' font-size was constant as 16px. That caused problems in the themes using smaller font sizes. For example Atom Dark is using 11px. Now the triangle icons are inheriting the font-size from the title.

Reviewed By: jamesgpearce

Differential Revision: D3903778

fbshipit-source-id: 2354039b5d5e90abd274986b2579c6baae93e10d