Diff View Cleanup: Fix restoring on reload and open with commit/publish contexts
Summary:
The restore on reload wasn't even working before the rewrite due to a wrong condition in the `serialize()` --> now fixed.
Also, with the rewrite, the diff view model is no longer in control of the view mode and commit mode switches - this diff delegates those to actions to be able to open the diff view to commit or publish mode from the file tree or editor context menu.

Reviewed By: jgebhardt

Differential Revision: D3901255

fbshipit-source-id: 1b62da3c97ae14c192c61c339546f72a0dc42c11