Diff View: Fix offsets initially not showing when switching from a different tab
Summary:
When switching from a different tab using the command, or the toolbar button, the editor is created, but not attached until the tab-switch is complete.
Hence, when setOffsets is called: editor line height = 0 --> so, the offsets are hidden.
Here, we refresh the offsets and call the editor element update callback only when the editor has completed attaching.

Reviewed By: jgebhardt

Differential Revision: D3903718

fbshipit-source-id: 9464ad577d18af66aaeda09b5ef155161b5d4346