Clang: Refactoring provider
Summary:
Based off nmote's draft provider API. Comes with a unit test that basically works.
This doesn't do formatting yet, but I think it's around to play around with.

I'll probably move the `refactor` logic to the server side so we don't have to pay 2 RPC requests to find references and then refactor.

Reviewed By: peterhal

Differential Revision: D3802856

fbshipit-source-id: 6fb79bdc655abe646357689003fdfed9412039eb