Remote Files: Display notifications when files are taking a long time to save/load
Summary:
With poor network connections, file saves may take a long time or even hang.

This is especially bad when people save on close, in which case they might close the editor before the save is done. When opening files, this just looks confusing.

To alleviate this, display a loading notification if the save/load is taking more than 500ms.
I think this is better than a busy signal, since this is such a critical piece of information (and I think it's easy to mistake a long save time for e.g. arc lint)

Changelog: Display a loading notification for slow remote file saves/loads.

Reviewed By: zertosh

Differential Revision: D3927486

fbshipit-source-id: 78b1f6255be93506eecbabbcfc861d8b13f9a336