Remove assert when disposing proxy after connection is closed.
Summary:
The remote-atom package disposes of the RemoteCommandService after
the remote connection is closed. This causes a (benign) assert when attempting
to send the dispose() over the rpc transport.

Silence that assert by not attempting to send the dispose message if the transport
is closed. Note that when the connection is closed, all proxies on the remote side
are dispose()ed by the rpc service.

Reviewed By: mostafaeweda

Differential Revision: D3864046

fbshipit-source-id: 9eaa92c3d17c76e5a4361baa4240f74368b0fcd1