Replace python eslint runner with js version
Summary: Originally I wrote this in python (see D3410807) because it was much faster to traverse the Nuclide dir for js files this way. Eslint has since fixed it's fs walking code, and this is now slightly faster than the python implementation. This also solves the issue of properly respecting the `.eslintignore`. In the python version, you couldn't have negations.

Reviewed By: hansonw

Differential Revision: D3937862

fbshipit-source-id: e07dae21f6da7ec806a26db31d966f01a2d96924