Dispose timers in launch attach
Summary:
I'm not sure /exactly/ what I did to trigger this, but I know it was some combination of toggling the debugger's
launch attach UI and shutting down the Nuclide server.  Whatever I did, it caused this setInterval timer
(normally used to refresh the attach process list) to not be cleared, and so every interval we would fail an
invariant since a debugger rpc service was null.

Poking around, I saw that the disposables for the object HAD been disposed, but the timerId was a number.  So
to stop this from happening again I'm adding the clearInterval into the disposable for the class.

Reviewed By: zertosh

Differential Revision: D3933401

fbshipit-source-id: f11a4ce8172e9ae7001d0c7cec8d053962236fc3