Display all Buck process messages until the first WebSocket message
Summary:
The Buck launcher often waits for the daemon or has to download the client.
When this happens, it emits output that we currently don't display.

Scrap the previous logic that buffered all messages, and just display all messages until the first socket message.

There's a slight risk that a delay in the socket message could lead to some extra output, but it's pretty unlikely in practice.

Changelog: Ensure that Buck loading messages are displayed in the console.

Reviewed By: matthewwithanm

Differential Revision: D3946595

fbshipit-source-id: e3f04b26fe54211adc72b33eb5ae93ae839a10a8