Ensure correct message ordering with HackProcess autocomplete.
Summary:
Before this diff, there was a race on a text edit event
between that edit getting to the HackProcess and the autocomplete request
getting to the hack process.

This Fixes that by adding HackProcess.getBufferAtVersion() which not only
waits for the correct TextBuffer version from the FileCache, but also waits for
the corresponding text edit to be sent to the hh_client process.

Reviewed By: mostafaeweda

Differential Revision: D3938220

fbshipit-source-id: ee07f78ba5a3607da0be37c50b68b7b9e7ae3eb3