Avoid context menu clashing with multiple instance of changed files tree
Summary: If multiple instances of the MultiRootChangedFilesTree are present then the context menu gets invoked from each of the instances. This happens because of the css selector being the same for all the instances. This change adds the commandPrefix to the className which would help us keep check on only the correct context menu being shown.

Reviewed By: mostafaeweda

Differential Revision: D3946178

fbshipit-source-id: 261bd788cfa2bb2d1a759983c39ae48a6ad68832