Process exitCode can be null.
Summary:
Process 'exit' events have both a code and a signal, exactly one of which
will be non-null. In particular killing a process yields a null code.
Handle this in our process library by using a ProcessExitMessage instead of a single
numeric code. Details are here: https://nodejs.org/api/child_process.html#child_process_event_exit

Use the new exitEventToMessage() to format a ProcessExitMessage in a way suitable for
display to a user.

Reviewed By: matthewwithanm

Differential Revision: D3955081

fbshipit-source-id: 40c051ac287dcf4b9a996026d94745710ee63b6e