Node wrapper around Python debugger.
Summary:
This introduces a new Python debugger with a JavaScript wrapper that is tailored
for Nuclide. As noted in the `README.md`, it may not be as full-featured as
other Python debuggers out there, but having our own makes it easier for us to
ensure that it works with both local and remote debugging with minimal setup in
Nuclide.

Reviewed By: johnislarry

Differential Revision: D3941993

fbshipit-source-id: dd1a48e8c44c44f82d83ff11a7b00bd9b14f08e3