Handle atom getPaths not being consistent
Summary:
In the uncommitted changes the repos currently active is fetched using atom getPaths method but on logging it is apparent that on removing
repos when there are multiple repos sometimes causes atom to assume the directory exists briefly even after it is removed. The first pass
causes the map to be cleared and the second tries to add it back but with an undefined value as the data is lost.

Reviewed By: mostafaeweda

Differential Revision: D3972891

fbshipit-source-id: 972cbfabdb72cf0673816d65d050cca23f20cdb7