Replace HHVM icon with "ARC"
Summary:
This used to be called the HHVM task runner. At some point, it was
renamed to "Arcanist," but the icon was never updated. Now we have an
actual HHVM task runner, so it's even weirder to use the HHVM icon.

Since Arcanist doesn't have an icon (and I think the Phabricator icon
would be misleading here), I just went with text.

Reviewed By: jgebhardt

Differential Revision: D3975652

fbshipit-source-id: e41098b164fe83f36d0bfa562ce12ec218bfa94b