Clang: fix edge case in diagnostic types
Summary:
There are cases where Clang diagnostic ranges can have empty paths.. apparently this can be triggered by adding/deleting headers without a full recompile.

Looks like we don't rely on this fact anywhere in the code, so this is a trivial change. It's just the RPC marshalling that threw.

Reviewed By: bolinfest

Differential Revision: D3978675

fbshipit-source-id: cdd68f68b93444cf5292e8101075b690d9ae7c5c