Check editor destroyed state before updating markers
Summary:
There's a few places in our code that add markers after the editor is destroyed, causing unhandled exceptions/rejections.

In the case of datatips and code-highlight, the editor events are debounced so they often trigger after editor destruction.

In the case of diagnostics, this is caused by a really subtle race where editor destruction triggers diagnostic invalidations, which triggers an editor update before the destroy callback cleans up.

Reviewed By: nmote

Differential Revision: D3983513

fbshipit-source-id: e070513dd40371960c5b6748cfd2aaddc29a4f8d