Show notification when waiting on remote file
Summary:
When the nuclide RPC mechanism is used to open a file locally from a remote server, it can sometimes get lost in the noise. If you are using a local nuclide as a remote editor, your terminal might be stalled while it waits for the file to be closed.

The feature is disabled by default behind a setting `shouldNotifyWhenRemoteServerIsWaitingOnFile`.
Closes https://github.com/facebook/nuclide/pull/763

Reviewed By: nmote

Differential Revision: D3982654

Pulled By: peterhal

fbshipit-source-id: b54a264d2e78564173774d8827c9bf5edf2067fb