iwdp debugger #5 -- send fake loader breakpoint pause event
Summary: Nuclide will auto-resume the first `Debugger.paused` event from the backend, because it assumes it's a loader breakpoint.  In this case, the target doesn't emit a loader breakpoint pause when we attach, so we'll just fake it.  If we don't do this, then the first time a breakpoint is triggered, it will be ignored, which is annoying.

Reviewed By: jgebhardt

Differential Revision: D3989723

fbshipit-source-id: 88b9c526c28ce9298ba7ac321720db5d5ef52250