Improve combobox sorting
Summary:
The current combobox doesn't have any preference on the order of the items to be displayed,
which is not the best when the are a lot of options and many with the same substrings.

In this diff I'm adding a small sorting function, first it sorts by index of match and then
by length of the result.

Reviewed By: jgebhardt

Differential Revision: D3991784

fbshipit-source-id: a8bca738b3a328aea5940e0d755456036e14c9da