Pass PaneContainer to Panel component
Summary: Originally, Panel displayed a single item and that could be any "viewable." Later, we started only putting PaneContainers in them. This makes it official, allowing us to get information about the things being displayed, and fixes a bug whereby `getPreferredWidth()`/`getPreferredHeight()` was called on the PaneContainer instead of its items.

Reviewed By: jgebhardt

Differential Revision: D3989816

fbshipit-source-id: 3abb032c1cdcad258346053d53161565232914c5