Buck: Throw an exception after Buck fails
Summary:
When the Buck process terminates with an error or non-zero exit code, throw.

This is a little tricky because I don't want to throw immediately and interrupt the diagnostics stream. Instead, in the side effect listener we'll save the error and throw it at the end (feels a little gross... open to other suggestions)

Reviewed By: matthewwithanm

Differential Revision: D3991803

fbshipit-source-id: c3dd77bb074fbefd0f81c32c3f8da27d9bdfe8ec