Diff View: Don't show the plain error message in the editor
Summary:
Previously, it was empty & that made some sense to mean the file doesn't exist at this revision.
Showing the error message was a bad idea, this reverts it.
In the future, the UI needs to be more clear about this file not existing at this revision, but error message parsing will be required.

Reviewed By: vjeux

Differential Revision: D3996428

fbshipit-source-id: 5b2d596a7138ff08205a46d24a4fd6900f82b96d