Handle and test for edge cases in refactoring
Summary:
It's not pleasant to have to have two separate places to handle unexpected
errors (the `.catch` on the root epic and the try/catch in the middleware). But
it works.

It's also not great to assert that there haven't been any uncaught errors the
way I'm doing it. But I haven't found a better alternative that allows me to
test this stringently.

Reviewed By: matthewwithanm

Differential Revision: D4005920

fbshipit-source-id: 609ed39291c2b4b71f4d9c24c45ce52870e99190