Remove custom checkbox styling
Summary:
Previously, Atom's checkbox styles were in settings-view so we had to duplicate them (in case settings-view was disabled). That's since been fixed but our duplicated styles were never removed, which became obvious when Atom updated theirs.

This commit removes most of our custom styling so that our checkboxes feel at home in Atom again.

The vertical alignment isn't perfect in the file tree, but it never was (and in fact, this is probably an improvement).

Reviewed By: jgebhardt

Differential Revision: D4013688

fbshipit-source-id: e638271e6c87c68a3da71f96a099ab3799af5cda