Enable/disable HHVM toolbar based on project root, not active file
Summary:
Script debugging is still toggled based on the extension of the active file.
However, the task should be enabled and disabled based on the active project.

This fixes a bug where the HHVM toolbar stays (misleadingly) enabled when a HHVM project is added and then removed.

Changelog: HHVM toolbar is now enabled/disabled based on the current working root.

Reviewed By: matthewwithanm

Differential Revision: D4024497

fbshipit-source-id: dfaa620683114a7a38ea62e621296949d52a163b