Allow optional properties to be undefined in the rpc framework
Summary:
This matches Flow. If you have

```
type Foo = {a?: string}
```

Flow will allow your object to be either `{}` or `{a: undefined}`.

Note that this is not a perfect solution because it does not preserve the
difference between those two objects. The deserialized object will simply omit
the property altogether. However, this is at least closer to the ideal, and it
prevents your Flow-typechecked code from getting a runtime error from the RPC
layer (in this particular case).

There is also precedence for this sort of behavior -- with the following type:

```
type Bar = {a: ?string}
```

the object `{a: undefined}` will also be deserialized as `{}`.

Reviewed By: hansonw

Differential Revision: D4024553

fbshipit-source-id: 5ebea8bd11e33d1a0995f7748625897afe870c5a