Change panel hiding behavior; allow empty panels to be shown
Summary:
Previously, empty panes could never be shown, however, this makes it difficult to add a toggle command without confusing behavior (i.e. nothing happening when you try to show an empty pane).

With this change, empty panes can be shown explicitly, but the pane will be hidden when the last item is removed. For most people, there should be no noticeable difference.

Reviewed By: vjeux

Differential Revision: D4024359

fbshipit-source-id: 4023b2463ad6a853ed29b3905a0b7fb9552c37dc