Use the require hook in Atom instead of patching their transpiler
Summary: The idea behind monkey patching Atom's transpiler was so that we could easily fallback to it if we needed to. Upgrading to babel 6 will introduce syntax that is unparseable by Atom's babel 5, so there's no benefit to monkey patching. Using the require hook is more straight-forward, and makes it easier to debug transpiler issues.

Reviewed By: hansonw

Differential Revision: D4025712

fbshipit-source-id: 866ca953cb0c1a79eab9bf73210f8d32bff87cd4