Killing process is not necessarily an error
Summary: An example of when killing a process is not an error is the `ios_webkit_debugger_proxy` binary I use to let the debugger communicate with the simulator.  This program is expected to run until killed.  When I unsubscribe from my observable subscription, this should kill the process (and does) but it's not an error.

Reviewed By: mostafaeweda

Differential Revision: D4034402

fbshipit-source-id: 92289ad2bc75ea7b699706976db5cfacd972e68c