Fix error catching in nuclide-ocaml
Summary: Catches errors from merlin in HyperclickProvider and shows them to the user. These errors are things like when the module you're trying to jump to hasn't been built, or your build artifacts are from the wrong version of ocaml, so I think it makes sense to raise them to the user.

Reviewed By: zertosh

Differential Revision: D4050324

fbshipit-source-id: f7d1b4317c560293c4ba85c4da5d8db0673a2a04