Make sure Scribe logging for unhandled exceptions works
Summary:
Some of our fatal server crashes aren't getting logged!

This is because `ScribeProcess.write` is async and therefore doesn't write until after the process closes.
While looking into this, I also realized that writes can potentially buffer, so `join` has to wait for the `drain` event if this is the case.

Finally, fix the mock to exit when stdin ends so that the test finishes much more quickly.

Reviewed By: shushz

Differential Revision: D4061457

fbshipit-source-id: 5e985b732b6e20b71a2b6d70aeee812c9e77d338