Add inline-invariant-tr
Summary:
This transform is half-optimization and half-convenience. Example:

```
// Before:
invariant(foo ? bar : baz);

// After:
if (!(foo ? bar : baz)) {
  throw new Error('Invariant violation: "foo ? bar : baz"');
}

 ---

// Before:
invariant(true === false, 'Always false');

// After:
if (!(true === false)) {
  throw new Error('Always false');
}
```

It converts `invariant()` into an `if` statement - which saves you a function call. And, when there is no error message, it inlines the test expression so the errors are maybe a bit more helpful when debugging.

Reviewed By: hansonw

Differential Revision: D4063126

fbshipit-source-id: 42e654dc8e1b1c3980e4925a07d81e8fff1d7be2