call ocamlmerlin -version to determine version
Summary:
old versions of ocamlmerlin don't support the `["protocol", "version"]` command, so nuclide throws an error (which `getVersion` caught, but it still logged about it unnecessarily, and could've caught other errors like merlin not being installed).

instead, we can call `ocamlmerlin -version` directly, which works in older versions too. it's not even an extra call, since we can avoid calling `which`.

I also changed the caching so that if merlin isn't found the first time, that result is cached instead of calling `which` again every time. Pretty sure this was running all the time when you mouse around an ocaml file. this just means that if you install merlin, you'll have to restart nuclide to pick it up. I think this is a fair tradeoff.

Reviewed By: peterhal

Differential Revision: D4053425

fbshipit-source-id: ee3c91c1ff3ce2524b8794e41d6fc14a01ca7228