Don't close files when they're renamed from the file tree
Summary:
Watchman sends renames as a delete + change, so we should handle this accordingly.

Since the file tree sets the path of the renamed buffer, we can detect this by flagging the delete event on the old file and re-subscribing to the new path.
(Note that we can't do this from `setPath`, as the rename hasn't actually taken place at that point).

This still doesn't work if you `mv` it from the file system, but that can't be easily solved right now.

Changelog: Don't close files when they're renamed from the file tree.

Reviewed By: matthewwithanm

Differential Revision: D4062912

fbshipit-source-id: c2c8851db74babb9914f4328c90261de1d4d7c41