Use SharedObservableCache for RemoteFile/RemoteDirectory watches
Summary:
The ServerConnection currently caches RemoteFile/RemoteDirectory objects, but this is very difficult and adds the overhead of having to subscribe to delete events to invalidate the cache.

However, the only expensive part of RemoteFile/Directory objects is the file watcher, so if we can cache that then we can just freely create files and directories. A follow-up diff removes the RemoteFile/Directory cache.

This is a use case where the factory must be provided at `SharedObservableCache.get`  time.

Reviewed By: shushz

Differential Revision: D4072057

fbshipit-source-id: 6bb5b757a50343ccaf6386a86380b7132f915999