Remove ServerConnection path/directory caching
Summary:
After RemoteFile/Directory objects cache their underlying file watchers, there is no need for these. This saves a tremendous amount of file watches since we no longer need to subscribe to every single one.

For example, if we expand a file tree directory, we used to subscribe to all subentities. Now, we only subscribe to the root.

Reviewed By: shushz

Differential Revision: D4072095

fbshipit-source-id: f9aaf6ebc727d2234c1a586cceedb0811cd7bd37