fix bad renaming
Summary:
It turns out that going from string literal to constants I replaced 'fetching' to VISIBLE. Oops.

In practice, the code kind of still works, but it's harder to make disappear. Now it's fixed.

I gave thoughts to adding tests to this class but i'm still not very sure how to. Ideas welcome.

Reviewed By: jgebhardt

Differential Revision: D4094213

fbshipit-source-id: f2bf9fdac8c2dc791adfb903f6464b3b74c2afb0