Move file-tree toolbar into the file-tree scroller (Atom 1.12 support)
Summary:
On Atom 1.12 the file-tree toolbar (the one that contains the Working Set icons and the search text) is rendered on the top of the sidebar section instead of being rendered near the tree that it actually controls.
From the inspection of the DOM tree I can't figure out why and how was it ever placed properly before 1.12. But anyway - this diff moves it to its place.

Reviewed By: jgebhardt

Differential Revision: D4092628

fbshipit-source-id: 5055900adcfe201fdea5cc294bb81188991b5e1e