Add Notiflyer component
Summary:
This adds a status bar enhancement that is currently unused. I'll be using it in the next diff.

The justification is that the contextual dialogue is often inappropriate for things like formatting error and jump-to-location errors. Either they stick or they get manually dismissed. They also block a big part of the screen and can't be swapped out in-place with an updated message of a different type (e.g. from error to info). Notiflyer makes it unobstructive and has been successfully used in the current atom-reason plugin externally.

As per the comments, I'll pull this out into an independent module later if this is popular enough.

{F64358542}

Reviewed By: matthewwithanm

Differential Revision: D4100961

fbshipit-source-id: c33c92419356e25814ada769686d6a690c816219