Diff View: Open on file change selection instead of just diff
Summary:
Previously, when the Diff View is a big complete chunk, it all comes alive together.
But now, the Navigator could come alive without the diff editors.

Then, clicking on the file changes should open the view if it's not already opened.
The easiest followed pattern, is to use Atom openers API: `atom.workspace.open` with the diff view url.

This diff takes this out to utils and shares the usage - but future cleanup could follow when the migration to the new UI is complete.

Reviewed By: aadisriram

Differential Revision: D4108566

fbshipit-source-id: aa7d0ee2af6ad1442e575066a2057f6481264d4b