Split Diff View: work-around Atom 1.12 broken block decorations
Summary:
In Atom 1.12.x, instead of deleting the removed block decorations, Atom adds a classname to it to hide it:
https://github.com/atom/atom/blob/v1.12.0-beta7/src/block-decorations-component.coffee#L75-L82
When switching between the tabs, I could see the old decorations that were destroyed - hence, this works around it.
Fortunately, this doesn't happen in Atom 1.13.x - so, we can remove this hack when we upgrade Atom.

Reviewed By: aadisriram

Differential Revision: D4128700

fbshipit-source-id: a514af25c2d2c4d524c0e1000785e272cf17ece2