Clang: strip compiler argument from compilation database commands
Summary:
Compilation databases contain the full compilation command, e.g. `g++ test.cpp <flag1> <flag2> ...`.
When we're giving flags to libclang, we don't actually want the first argument (the compiler), and in fact this sporadically causes a "linker error: input unused".

I can't repro it consistently but I've seen it happen before - this appears to be the cause.

Reviewed By: nmote

Differential Revision: D4133293

fbshipit-source-id: 15b6df59587a8a1ae83acf016cbc909cbdb883f2