Dont throw when attempting to close a file after a reconnect.
Summary:
This happens when the network goes down after the file close
has been sent to the server, but before the response has been recieved.

The server will record the file as closed. The call from the client will timeout.
The client will continue to think the file is open, and will continue to resend the
close.

Reviewed By: hansonw

Differential Revision: D4134991

fbshipit-source-id: 4740ec6b7e871f8c7bdcbdfbe884e69653d5e57e