Create Nuclide urlMain hook to handle atom://nuclide URLs
Summary:
This enables atom://nuclide/path?param=1 URLs to be intercepted and processed from within Nuclide.

https://github.com/atom/atom/pull/11399 is an open PR on the Atom side, but likely requires some work and it'll be a while before this is ready. I plan to work on a new proposal to drive that forward, but in the meantime this leverages the existing `urlMain` functionality that Atom provides to make it work.

A quick rundown of how this works:

- Atom provides a `package.json`-level `urlMain` hook. If provided, `atom://{package}` URIs launch your `urlMain` code rather than the standard window initialization script: https://github.com/atom/atom/blob/master/src/main-process/atom-application.coffee#L627
- This adds a `urlMain` hook for Nuclide. There are two cases to handle:

This comes with tests to ensure that the core functionality continues to work across Atom versions.
It's by no means complete, but I imagine it should hold us over until we can upstream a saner method.

Reviewed By: zertosh

Differential Revision: D4085694

fbshipit-source-id: b4f091cf53d2e0d04d5cb1f6d13fbc38da2d3846