Never leave PHP debugger connections in a starting state without resuming
Summary:
Well this is a "fun" one.

The XDebug protocol expects a 'run' command sent when the connection is in 'starting' mode.  Prior to this diff, there were some cases where we would never send this 'run' command, and the connection would be left in 'starting' mode.  This manifested itself (in some instances) as PHP scripts hanging, because Nuclide's debugger would never allow execution to begin.

This would only happen under certain circumstances when the "Stop one stop all" checkbox was checked.  If this feature is turned on, then pausing one request in Nuclide's debugger will pause every request.  This bug is an edge case of this feature.

Big thanks to sid0 and paulbiss for doing almost all of the leg work to troubleshoot this tricky subtle issue!

Reviewed By: yinghuitan

Differential Revision: D4129088

fbshipit-source-id: 315d81b11e5ad77026cf11121f5c3be7c58bebd0