Use performance.now when available
Summary:
performance.now has some additional precision. Also, we want to round at the very last time in order to avoid losing precision.

Note: this diff doesn't really matter because the stats we are looking at are hundreds of ms to seconds, so sub-millisecond differences isn't very important. But I figured that I would make it a bit more correct :)

Reviewed By: zertosh

Differential Revision: D4148610

fbshipit-source-id: c19166b0b91c6b448a60fbfa37176a2ac55cfb15