Add list of external process called and their duration
Summary:
During my performance investigations, the first thing I do is always to print all the external processes that are executed and the time they take. I figured out that it was pretty simple to log this information and display it in the health panel.

Thanks to this, when we get a report of <x> is slow, we'll be able to tell them to open their health panel and see all the commands + time it took.

Note: this doesn't take into account remotely executed commands.

Reviewed By: zertosh

Differential Revision: D4150801

fbshipit-source-id: a3ac0172e1b251c6e03d270b7304862e7dd1e911