Unbreak open source
Summary:
I didn't realize that I was importing things from fb/ folder within open-sourced files. Thanks zertosh for the heads up.

In order to fix it, I create a new Epics.js file in the fb/ folder and changed the type of suggestedReviewers from Array<FbInternUser> to Array<string>. I didn't even use in my code any of the FbInternUser stuff so it was wasteful anyway.

Reviewed By: zertosh

Differential Revision: D4167285

fbshipit-source-id: 4f5a618fb4d7429e8f62e9daa73434e64f8c933a