Refactor context menu item acts on right-click location
Summary:
Without this hack, it would act on the location where the cursor was, instead of
where the user right-clicked.

This isn't great: in particular, things will get weird if people run the command
themselves without going through the context menu. But I had the Atom folks look
at it and they said it seemed like the best way to address this problem.

Reviewed By: vjeux

Differential Revision: D4169278

fbshipit-source-id: 59e74fb292c8daa388d2c214b42f36fcb574bfee