Bypass flow wrapper
Summary:
The shell wrapper adds ~20-25ms of delay for every flow command. This is particularly important for autocomplete and hyperclick.

We already had all the infrastructure in place to find where flow is and periodically invalidate that cache, so it's just one-time indirection at startup to win on every single subsequent ones :)

Reviewed By: nmote

Differential Revision: D4161759

fbshipit-source-id: 02c5bfead2df47a3050dce21dac457dd628f3aa9