Improve task runner initialization story
Summary:
This commit reworks task runner initialization. Instead of displaying the toolbar immediately, we wait until all task runners are "ready" (defined as having emitted their first task list). We then evaluate which of those tasks should be selected initially and, finally, show the toolbar. If the project root changes, we re-evaluate an initial task (but leave visibility untouched).

There are a lot of tests in this diff in order to account for a lot of edge cases. However, the current Atom package serialization behavior is going to require further tweaks to get this to where we want it to be. Nuclide features are still being serialized after others have deactivated, leaving aggregator packages unable to serialize aggregated information, depending on the order of deactivation at shutdown.

Also to do yet: Make sure the test runners don't emit disabled tasks immediately (since those won't be prioritized for the initial task).

Reviewed By: hansonw

Differential Revision: D4156747

fbshipit-source-id: a19d3a790b20d7690acfcedb6a980d17d3f8477b