Add support for selecting text
Summary:
matthewwithanm told me that someone mentioned that they couldn't copy and paste things from the debugger pane.

Since I did a similar thing for the health tab in D4150558, I figured I would do the same for the debugger.

This is pretty ghetto as it doesn't interrupt all the click handlers so you need to cmd-c before mousing up, but at least you can do it, so it's better than before.

Reviewed By: matthewwithanm

Differential Revision: D4151060

fbshipit-source-id: 0f0c20bad8df924e48799cb57ecce150cd515ee9