Shift all the revisions up by one
Summary:
I've been very confused by the revision UI. It is not intuitive to me that when you click on a commit, it doesn't show the associated changes.

What this diff does:
- removing the last commit which is one from some random person you happened to be rebased on
- be able to click on Uncommitted changes to see them
- clicking on a commit shows the file changes of the commit itself as well

I personally think it's a better experience but I'm not sure if I think differently than most people on this point. Would love some feedback :)

Reviewed By: mostafaeweda

Differential Revision: D4172010

fbshipit-source-id: e545f34d6d04387b0006b10edca4dbf687f62242