Remove options from featureConfig.observe
Summary:
I wanted to change the type of the callback to be more strict (use `mixed`
instead of `any`), but when I changed it, I got no Flow errors because `Object`
also matches any function. We weren't using the options object anyway, and it
looks pretty arcane, so this seems like a good choice.

Reviewed By: vjeux

Differential Revision: D4186493

fbshipit-source-id: 44259af1838e66d04f25287c50ec986290522929