Don't serialize Nuclide features during the deactivation phase
Summary:
Atom [[ https://github.com/atom/atom/pull/12759 | is no longer serializing ]] as part of the deactivation phase when reloading/shutting down. That's good. But UP's `deactivate()` calls back into `atom.packages.deactivatePackage()` causing it to serialize.

This commit makes it so that we never serialize as part of our deactivation. However, this change alone would mean that disabling the Nuclide package wouldn't cause any serialization (since the Nuclide package has no `serialize()`). Therefore, the commit also adds a `serialize()` method. Unfortunately, this has the unwanted side-effect of serializing all of the Nuclide features twice when reloading/quitting Atom: once because of Nuclide's `serialize()` and once because of the package's `serialize()`. The performance cost of this is negligible (~2ms).

Reviewed By: zertosh

Differential Revision: D4186146

fbshipit-source-id: de51274976dc1298d6fd2da9482c82eab2ccf37a